--- GraphSON/readme.md
@@ -1,7 +1,0 @@
-﻿# GraphSON
-
-A library containing definitions for GraphSON edges and vertices. 
-
-## Vertex Deserialization
-
-If a vertex was serialized using the [Query](../Query) packate, it can be deserialized by calling the `Vertex.Deserialize<objecttype>()` method.
