--- UnitTests/UnitTests.GraphSON/UnitTests.GraphSON.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-
-    <Version>0.1.0</Version>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Gremlin.Net" Version="3.3.2-rc1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\GraphSON\GraphSON.csproj" />
-  </ItemGroup>
-
-</Project>
